Updated tests as promised
