+++ apps/client/src/components/Modal.test.js
@@ -3,13 +3,15 @@ import React from "react";
 import {
   MODAL,
   MODAL_CLOSE_BUTTON,
+  MODAL_DECLINE_BUTTON,
   MODAL_OPEN_BUTTON,
 } from "../utils/test-ids";
 import { act, cleanup, fireEvent, render } from "../utils/test-utils";
 import Modal from "./Modal";
 
+window.scrollTo = jest.fn();
+
 const onOpenMock = jest.fn();
-const onConfirmMock = jest.fn();
 
 afterEach(cleanup);
 
@@ -23,9 +25,10 @@ it("Modal Button should render, but the Modal itself not", () => {
   expect(queryByTestId(MODAL)).not.toBeInTheDocument();
 });
 
-it("Modal should open and close", () => {
+it("Modal should open and close", async () => {
   const { getByText, queryByTestId } = render(
     <Modal
+      backdropOpacity={20}
       openButtonText="button"
       heading="heading"
       handleOpenModal={onOpenMock}
@@ -34,6 +37,7 @@ it("Modal should open and close", () => {
     </Modal>
   );
 
+  // Modal = open
   act(() => {
     fireEvent.click(getByText("button"));
   });
@@ -45,36 +49,22 @@ it("Modal should open and close", () => {
   act(() => {
     fireEvent.click(queryByTestId(MODAL_CLOSE_BUTTON));
   });
+  // Modal = closed
 
   expect(queryByTestId(MODAL)).not.toBeInTheDocument();
-});
-
-// @TODO: finish this test
-xit("Modal should handle the footer buttons", () => {
-  const { getByText, queryByTestId } = render(
-    <Modal
-      openButtonText="button"
-      handleConfirmButton={onConfirmMock}
-      showCloseButton
-    >
-      text
-    </Modal>
-  );
 
   act(() => {
     fireEvent.click(getByText("button"));
   });
+  // Modal = open
 
   expect(queryByTestId(MODAL)).toBeInTheDocument();
-  expect(getByText("heading")).toBeInTheDocument();
-  expect(onOpenMock).toHaveBeenCalledTimes(1);
+  expect(onOpenMock).toHaveBeenCalledTimes(2);
 
   act(() => {
-    fireEvent.click(queryByTestId(MODAL_CLOSE_BUTTON));
+    fireEvent.click(queryByTestId(MODAL_DECLINE_BUTTON));
   });
+  // Modal = closed
 
   expect(queryByTestId(MODAL)).not.toBeInTheDocument();
 });
-
-// Lines to check:
-// 88,123,124,132
