+++ apps/client/src/utils/test-ids.js
@@ -24,5 +24,6 @@ export const STEPBYSTEPITEM = "step-by-step-item";
 
 export const MODAL = "modal";
 export const MODAL_CLOSE_BUTTON = "modal-close-button";
+export const MODAL_DECLINE_BUTTON = "modal-decline-button";
 export const MODAL_CONFIRM_BUTTON = "modal-confirm-button";
 export const MODAL_OPEN_BUTTON = "modal-open-button";
